+++ internal/mgtt-plugins/auth/onAuthUsersListGet.go
@@ -3,14 +3,17 @@ package auth
 import (
 	"encoding/json"
 
+	"github.com/rs/zerolog/log"
 	"gitlab.com/mgtt/internal/mgtt/broker"
+	"gitlab.com/mgtt/internal/mgtt/clientlist"
 )
 
 type userListElement struct {
 	Username string `json:"username"`
 }
 
-func onHandleUserList(originClientID string) {
+func onAuthUsersListGet(originClientID string) {
+	var err error
 	// a new list
 	var newUserList []userListElement
 
@@ -25,7 +28,11 @@ func onHandleUserList(originClientID string) {
 	jsonData, err := json.Marshal(newUserList)
 	if err == nil {
 		if broker.Current != nil {
-			broker.Current.PublishToClient(originClientID, "$SYS/auth/users/list/json", jsonData)
+			err = clientlist.PublishToClient(originClientID, "$SYS/auth/users/list/json", jsonData)
 		}
+	}
+
+	if err != nil {
+		log.Error().Err(err).Send()
 	}
 }