+++ internal/mgtt/broker/handleNewClient.go
@@ -14,8 +14,9 @@ func handleNewClient(newConnection net.Conn) {
 
 	var err error
 	var recvdPacket packets.ControlPacket
+	var newClient *client.MgttClient = &client.MgttClient{}
 
-	newClient := client.New(newConnection, ConnectTimeout)
+	newClient.Init(newConnection, ConnectTimeout)
 	err = clientlist.Add(newClient)
 
 	if err == nil {
@@ -55,6 +56,9 @@ func handleNewClient(newConnection net.Conn) {
 
 		// Remove the client from the list
 		clientlist.Remove(newClient.ID())
+
+		// inform our plugins
+		plugin.CallOnDisconnected(newClient.ID())
 
 	}
 }