+++ internal/mgtt/clientlist/Client.go
@@ -1,0 +1,45 @@
+package clientlist
+
+import (
+	"net"
+
+	"github.com/eclipse/paho.mqtt.golang/packets"
+)
+
+// Client  represents a client
+type Client interface {
+	Init(connection net.Conn, secondsTimeout int64)
+	ResetTimeout()
+	Close() (err error)
+	RemoteAddr() (remoteAddr string) // return the remoteAddr as string
+
+	// ID's
+	IDSet(id string)
+	ID() string
+
+	UsernameSet(username string)
+	Username() string
+
+	// Last-will
+	LastWillSet(packet *packets.PublishPacket)
+	LastWillGet() (packet *packets.PublishPacket)
+
+	// packet handling
+	PacketRead() (packet packets.ControlPacket, err error) // reads a single packet from the connection
+	Publish(packet *packets.PublishPacket) (published bool, err error)
+
+	// subscriptions
+	SubScriptionAdd(topic string)
+	SubScriptionsAdd(topics []string)
+
+	//
+	SendConnack(ReturnCode byte) (err error)
+	SendConnect(username, password, clientid string)
+	SendPingreq()
+	SendPingresp()
+	SendPuback(MessageID uint16) (err error)
+	SendPubcomp(MessageID uint16) (err error)
+	SendPubrec(MessageID uint16) (err error)
+	SendPubrel(MessageID uint16) (err error)
+	SendSuback(MessageID uint16, ReturnCodes []byte) (err error)
+}