+++ internal/mgtt/plugin/CallOnAcceptNewClient.go
@@ -14,6 +14,7 @@ func CallOnAcceptNewClient(clientID string, username string, password string) (a
 		if plugin.OnAcceptNewClient != nil {
 			log.Debug().Str("plugin", pluginName).Msg("call OnAcceptNewClient")
 			accepted = accepted && plugin.OnAcceptNewClient(clientID, username, password)
+			log.Debug().Str("plugin", pluginName).Bool("accepted", accepted).Send()
 		}
 	}